Refactor the custom_pg test into an example
Custom Pg types can interfere with applying migration to other
database backends (see #320), meaning the `Point` type shouldn't be
built at all when sqlite is enabled. To make this easier to manage,
move it out into an example rather than a test under the main `butane` library.
