+++ examples/custom_pg/README.md
@@ -1,0 +1,7 @@
+# Butane Custom Pg Example
+
+Demonstrates Butane's support for defining field types (i.e. types
+that may be used as fields in Butane models) which are represented by
+complex/nonstandard Postgres types at the database layer.
+
+Currently set up to run only as a test.
